requirements.txt
asgiref==3.7.2
Django~=4.1.0
djangorestframework==3.14.0
django-filter==23.2

PyYAML==6.0.1
uritemplate==4.1.1

# for visual studio code
autopep8
pylint
pylint-django
pylint-plugin-utils
flake8
pyflakes
